+++ tests/test_standard_deviation_box.py
@@ -30,7 +30,7 @@ def test_standard_deviation_box():
         )
     )
 
-    result = cle.create(test1)
+    result = cle.create(test1, dtype=float)
     cle.standard_deviation_box(test1, result, 1, 1, 0)
 
     a = cle.pull(result)
