+++ tests/test_variance_sphere.py
@@ -30,7 +30,7 @@ def test_variance_sphere():
         )
     )
 
-    result = cle.create(test1)
+    result = cle.create(test1, dtype=float)
     cle.variance_sphere(test1, result, 1, 1, 0)
 
     a = cle.pull(result)
